split off part of large file
