+++ enclone_main/src/lib.rs
@@ -10,6 +10,7 @@ pub mod flag_defective;
 pub mod inconsistent;
 pub mod main_enclone;
 pub mod merge_onesies;
+pub mod opt_d_val;
 pub mod populate_features;
 pub mod sec_mem;
 pub mod setup;
